requirements.txt
chainlit==0.7.700
openai==1.3.5
cohere==4.37
python-dotenv==1.0.0 
tiktoken==0.5.1
pinecone-client==2.2.4
langchain==0.0.350

youtube-search-python==1.6.6
youtube-transcript-api==0.6.1
langchain-community==0.0.3
scrapetube==2.5.1